Implement own version of withApollo at app level
