+++ global.d.ts
@@ -1,0 +1,5 @@
+declare namespace NodeJS {
+  interface Process {
+    browser: boolean;
+  }
+}
